Moved loginPage to its own file
