--- frontend/src/App.css
@@ -1,79 +1,0 @@
-body {
-  font-family: Arial, sans-serif;
-  background-color: #f4f4f4;
-  margin: 0;
-  padding: 0;
-}
-
-.login-container {
-  width: 300px;
-  margin: 100px auto;
-  padding: 20px;
-  background-color: #fff;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  border-radius: 5px;
-}
-
-h2 {
-  text-align: center;
-  margin-bottom: 20px;
-}
-
-.form-group {
-  margin-bottom: 15px;
-}
-
-label {
-  display: block;
-  margin-bottom: 5px;
-}
-
-input[type="text"],
-input[type="password"] {
-  width: 90%;
-  padding: 10px;
-  border: 1px solid #ccc;
-  border-radius: 3px;
-}
-
-.submit {
-  width: 100%;
-  padding: 10px;
-  background-color: #28a745;
-  color: white;
-  border: none;
-  border-radius: 3px;
-  cursor: pointer;
-}
-
-.submit:hover {
-  background-color: #218838;
-}
-
-.google-login {
-  margin-top: 3%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: rgb(232, 17, 17);
-  border: 1px solid #ddd;
-  color: #ffffff;
-  font-size: 14px;
-  font-weight: 500;
-  padding: 10px 15px;
-  border-radius: 5px;
-  cursor: pointer;
-  transition: background-color 0.3s ease, box-shadow 0.3s ease;
-  width: 100%;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.google-login:hover {
-  background-color: #7f0404;
-}
-
-.error {
-  color: red;
-  text-align: center;
-  margin-top: 10px;
-}
